backend/requirements.txt
# Core web framework
fastapi>=0.116.0
uvicorn[standard]>=0.30.0

# Data validation
pydantic>=2.11.0

# Environment variables
python-dotenv>=1.0.0

# NLP Libraries - Version aligned with your downloaded models
spacy>=3.8.0,<3.9.0

# Machine Learning Libraries
transformers>=4.44,<5
torch>=2.1,<3

# Optional, but good to have
httpx>=0.27.0